backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Vector Database - Qdrant
qdrant-client==1.7.0

# AI/LLM
openai==1.3.7
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# Embeddings & NLP
sentence-transformers==2.2.2
tiktoken==0.5.2

# Markdown & Frontmatter
python-frontmatter==1.0.1

# HTTP & Async
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For TestClient

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.0
mypy==1.7.1

# Type Stubs
types-python-dateutil==2.8.19.14
